% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_adagrad.R
\name{my_adagrad}
\alias{my_adagrad}
\title{Adagrad for regression}
\usage{
my_adagrad(X, Y, lr, beta_0, full)
}
\arguments{
\item{X}{An n x p matrix of predictors where rows are observations and columns are predictors.}

\item{Y}{An n x 1 vector quantitative response variable.}

\item{lr}{A constant that is the learning rate.}

\item{beta_0}{An p x 1 vector that is the initialization for the coefficients.}

\item{full}{boolean, if true uses the full G matrix for the update step, otherwise uses only the diagonal elements of G.}
}
\value{
An n x p matrix where each ith row is the coefficients for the ith iteration and the columns are predictors.
}
\description{
Adaptive gradient descent for linear regression.
}
\examples{
my_adagrad(X=X, Y=Y, lr=0.00001, beta_0=rep(0, ncol(X)), full=T)
my_adagrad(X=X, Y=Y, lr=0.00001, beta_0=runif(ncol(X)), full=F)

}
