% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_adam.R
\name{my_adam}
\alias{my_adam}
\title{Adam for regression}
\usage{
my_adam(X, Y, lr, beta_0, rho_1, rho_2, epsilon, regression = T)
}
\arguments{
\item{X}{An n x p matrix of predictors where rows are observations and columns are predictors.}

\item{Y}{An n x 1 vector quantitative response variable.}

\item{lr}{A constant that is the learning rate.}

\item{beta_0}{An p x 1 vector that is the initialization for the coefficients.}

\item{rho_1}{constant for 1st moment decay rate}

\item{rho_2}{constant for 2nd moment decay rate}

\item{epsilon}{positive constant for nonzero/invertibility}

\item{regression}{Boolean, regression if true, else classification.}
}
\value{
List where first elemnt is an n x p matrix where each ith row is the coefficients for the ith iteration and the columns are predictors and second element is a nx1 vector of runtimes for each iteration.
}
\description{
Adaptive moment estimation for linear regression.
}
\examples{
my_adam(X=X, Y=Y, lr=0.0000001, beta_0=rep(0, p), rho_1=0.9, rho_2=0.999, epsilon=1e-8)

}
