% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_OGD.R
\name{my_OGD}
\alias{my_OGD}
\title{OGD for regression}
\usage{
my_OGD(X, Y, lr = 1e-04, beta_0 = rep(0, ncol(X)), regression = T)
}
\arguments{
\item{X}{An n x p matrix of predictors where rows are observations and columns are predictors.}

\item{Y}{An n x 1 vector quantitative response variable.}

\item{lr}{A constant that is the learning rate.}

\item{beta_0}{An p x 1 vector that is the initialization for the coefficients.}

\item{regression}{Boolean, regression if true, else classification.}
}
\value{
List where first elemnt is an n x p matrix where each ith row is the coefficients for the ith iteration and the columns are predictors and second element is a nx1 vector of runtimes for each iteration.
}
\description{
Online gradient descent for linear regression.
}
